% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bbox2polys.R
\name{bbox2polys}
\alias{bbox2polys}
\title{create polygons from dataframe of bounding coordinates and ids}
\usage{
bbox2polys(idxxyy)
}
\arguments{
\item{idxxyy}{dataframe with 5 columns - id, minx,maxx, miny, maxy - representing the id, and bounding x,y of polygons}
}
\value{
SpatialPolygons object
}
\description{
create polygons from dataframe of bounding coordinates and ids
}
\details{
create polygons from dataframe of bounding coordinates and ids


 This program is free software but it is provided WITHOUT WARRANTY
 and with ABSOLUTELY NO GUARANTEE of fitness or functionality for any purpose;
 you can redistribute it and/or modify it under the terms of the GNU
 General Public License as published by the Free Software Foundation;
 either version 2 of the License, or (at your option) any later version.

\cr
Revision History
\tabular{ll}{
1.0 \tab 2017 March 08 Created\cr
}
}
\examples{
 #none yet

}
\seealso{
\code{\link{scan_las}}\cr \code{\link{read_las}}\cr
}
\author{
Jacob Strunk <Jstrunk@fs.fed.us>
}
