% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/catalog1.R
\name{catalog1}
\alias{catalog1}
\title{tweak to lidR::catalog that enables file read failure}
\usage{
catalog1(
  folder,
  progress = TRUE,
  select = "*",
  filter = "",
  chunk_size = 0,
  chunk_buffer = 30,
  verbose = FALSE,
  ...
)
}
\arguments{
\item{folder}{?}

\item{progress}{?}

\item{select}{?}

\item{filter}{?}

\item{chunk_size}{?}

\item{chunk_buffer}{?}

\item{...}{?}
}
\value{
<Delete and Replace>
}
\description{
tweak to lidR::catalog that enables file read failure
}
\details{
(Whatever license that JR has on lidR or the following)
 This program is free software but it is provided WITHOUT WARRANTY
 and with ABSOLUTELY NO GUARANTEE of fitness or functionality for any purpose;
 you can redistribute it and/or modify it under the terms of the GNU
 General Public License as published by the Free Software Foundation;
 either version 2 of the License, or (at your option) any later version.

\cr
Revision History
\tabular{ll}{
1.0 \tab 2024/03/28 started \cr
1.1 \tab date and revisions.. \cr
}
}
\examples{
 <Delete and Replace>


}
\author{
Jean Romaine and borrowed / modified by Jacob Strunk <something@something.com>
}
