% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/clip_plots1.R
\name{clip_plots1}
\alias{clip_plots1}
\title{clip plots or polygons form lidar}
\usage{
clip_plots1(
  polys = NA,
  poly_id = "id",
  dir_las = NA,
  dir_dtm = NA,
  file_las = NA,
  file_dtm = NA,
  pattern_las = c("[.]las$", "[.]laz$")[1],
  pattern_dtm = c("[.]img$", "[.]tif$", "[.]dtm$", "[.]vrt$")[1],
  dir_out = NA,
  height = T,
  n_core = 6,
  skip_existing = T,
  fix_dsm_bug = F,
  recurse_las = F,
  recurse_dtm = F,
  return = T,
  temp = "c/temp/clip_plots/"
)
}
\arguments{
\item{dir_las}{in case path to las has changed from project}

\item{dir_dtm}{in case path to dtms has changed from project}

\item{dir_out}{where to send clipped plots}

\item{height}{T/F subtract the ground}

\item{n_core}{number of cores to use in clipping}

\item{skip_existing}{skip csv files that have been processed already - in case processing was interupted}

\item{return}{T/F return intersectiosn}

\item{plot_polys}{sf polygons for plots}

\item{id_field_plots}{(optional) in the event that a polygon is provided}

\item{project_gpkg}{spatial polygons dataframe (intersection between dtm and las tiles)}

\item{project}{path to spatial polygons dataframe (intersection between dtm and las tiles)}

\item{plot_tile_intersect}{path to spatial polygons dataframe representing intersection between plots and project}

\item{do_plot}{T/F make a plot of the intersections}
}
\value{
<Delete and Replace>
}
\description{
clip plots or polygons form lidar - requires that one first make a project using project which
 includes both lidar and dtms
}
\details{
There are three ways to provide the input shapes to this function - either as a
 list of plot ids, coordinates and plot diameters (4 column data.frame), as a list of vertices
 organized by plot_id (3 column data.frame), or as a polygon shapefile (path to shapefile) inwich case
 the plot id field must be specified "id_field_plots".


 This program is free software but it is provided WITHOUT WARRANTY
 and with ABSOLUTELY NO GUARANTEE of fitness or functionality for any purpose;
 you can redistribute it and/or modify it under the terms of the GNU
 General Public License as published by the Free Software Foundation;
 either version 2 of the License, or (at your option) any later version.

\cr
Revision History
\tabular{ll}{
1.1 \tab move to SF objects and geopackages \cr
}
}
\examples{
 <Delete and Replace>

}
\seealso{
\code{\link{another_function}}\cr \code{\link{yet_another_function}}\cr
}
\author{
Jacob Strunk <Jstrunk@fs.fed.us>
}
