% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/project_create.R
\name{project_create}
\alias{project_create}
\title{Build point cloud processing tiling project (.laz, .las, .dtm) for use with FUSION}
\usage{
project_create(
  dir_las = NA,
  dir_dtm = NA,
  recurse_dtm = F,
  recurse_las = F,
  dir_out = "c:/lidar_projects/",
  layer_project = "RSForInvt_prj",
  layer_config = "RSForInvt_config",
  overwrite_project = T,
  project_dtm = "someProject_dtm",
  project_las = "someProject_las",
  dtm_year = "2099",
  las_year = "2099",
  do_scan_dtms = T,
  do_scan_las = T,
  duplicate_las = c("ignore", "remove"),
  duplicate_dtm = c("ignore", "remove"),
  tile_size = 1650,
  pixel_size = 66,
  xmn = c(NA, 561066),
  xmx = c(NA, 2805066),
  ymn = c(NA, 33066),
  ymx = c(NA, 1551066),
  proj4 = c(NA,
    "+proj=lcc +lat_1=47.33333333333334 +lat_2=45.83333333333334 +lat_0=45.33333333333334 +lon_0=-120.5 +x_0=500000.0001016001 +y_0=0 +ellps=GRS80 +towgs84=0,0,0,0,0,0,0 +units=us-ft +no_defs"),
  mask = NA,
  return = T
)
}
\arguments{
\item{dir_las}{where are las files}

\item{dir_dtm}{where are FUSION dtm files - eventuall enable any dtm type (.img, .tif etc)}

\item{project_dtm}{dtm project name}

\item{project_las}{las project name}

\item{dtm_year}{year of dtm files}

\item{las_year}{year of las files}

\item{tile_size}{processing tile size}

\item{pixel_size}{raster pixel size}

\item{xmn, xmx, ymn, ymx}{bound box for processing grid}

\item{dir_project}{where to place project}

\item{project}{project name}

\item{scan_dtms}{?scan dtm files}

\item{scan_las}{?scan las files}

\item{crs}{projection string}
}
\value{
<Delete and Replace>
}
\description{
Scans lidar and dems and finds out where they intersect a project tiling scheme
}
\details{
<Delete and Replace>

\cr
Revision History
\tabular{ll}{
1.0 \tab 5/07/2020 New package derived from old lasR package \cr
}
}
\examples{

proj_tn = RSForInvt::project_create(
  #'proj = project_create(
  dir_las="D:\\\\Box\\\\VMARS\\\\Projects\\\\DAP_evaluation_Meston\\\\Data\\\\Tennessee\\\\lidar_tiles\\\\"
  ,dir_dtm="D:\\\\Box\\\\VMARS\\\\Projects\\\\DAP_evaluation_Meston\\\\Data\\\\Tennessee\\\\DTM_fusion\\\\"
  ,path_gpkg_out="D:\\\\Box\\\\VMARS\\\\Projects\\\\DAP_evaluation_Meston\\\\R\\\\DAP_Lidar_analysis\\\\RSForInvt\\\\project\\\\TNLidar_RSForInvtProject.gpkg"
  ,layer_project = "RSForInvt_prj"
  ,layer_config = "RSForInvt_config"
  ,overwrite_project = T
  ,project_dtm="lidar_dtm"
  ,project_las="lidar_las"
  ,dtm_year="2018"
  ,las_year="2018"
  ,do_scan_dtms=F #'we already scanned the dtm folder - ok, to rescan, but slower
  ,do_scan_las=F #'we already scanned the las folder - ok, to rescan, but slower
  ,duplicate_las = c("ignore","remove")
  ,duplicate_dtm = c("ignore","remove")
  ,tile_size=1650
  ,pixel_size=66
  #',xmn=561066,xmx=2805066,ymn=33066,ymx=1551066
  ,proj4 = "+proj=lcc +lat_1=47.33333333333334 +lat_2=45.83333333333334 +lat_0=45.33333333333334 +lon_0=-120.5 +x_0=500000.0001016001 +y_0=0 +ellps=GRS80 +towgs84=0,0,0,0,0,0,0 +units=us-ft +no_defs"
  ,mask=NA
  ,return=T
)


}
\seealso{
\code{\link{scan_dtm}}\cr \code{\link{scan_las}}\cr


desired updates:
 add ability to scan fusion dtms and generic raster dtms
}
\author{
Jacob Strunk <Jstrunk@fs.fed.us>
}
