% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/raster2FUSION.R
\name{raster2FUSION}
\alias{raster2FUSION}
\title{<Delete and Replace>}
\usage{
raster2FUSION(
  dir_in,
  dir_out,
  pattern = "[.]img",
  recursive = T,
  ncore = 4,
  ...
)
}
\arguments{
\item{dir_in}{where are the input rasters}

\item{dir_out}{where to put output .dtm fusion rasters}

\item{pattern}{pattern for seeking rasters in dir_in}

\item{recursive}{should function recurse into subfolders}

\item{ncore}{use more than one core to process ?}

\item{...}{other arguments to list.files}
}
\value{
<Delete and Replace>
}
\description{
<Delete and Replace>
}
\details{
<Delete and Replace>

\cr
Revision History
\tabular{ll}{
1.0 \tab date and revisions.. \cr
1.1 \tab date and revisions.. \cr
}
}
\examples{
 <Delete and Replace>

}
\seealso{
\code{\link{list.files}}\cr \code{\link{raster}}\cr
}
\author{
Jacob Strunk <Jacob.strunk@usda.gov>
}
