% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_header.R
\name{read_header}
\alias{read_header}
\title{read an las/laz file header}
\usage{
read_header(paths = NA, dir_las = NA, recursive = F, pattern = "[.]la.$", ...)
}
\arguments{
\item{paths}{vector of paths to las files}

\item{dir_las}{folder with las files in it}

\item{recursive}{recurse into subdirectories}

\item{pattern}{pattern to find las/laz files}

\item{...}{other arguments to list.files}
}
\value{
<Delete and Replace>
}
\description{
read an las/laz file header
}
\details{
read an las/laz file header

\cr
Revision History
\tabular{ll}{
1.0 \tab 2020 May 07 Split from read_las \cr
}
}
\examples{
   files=list.files("C:\\\\temp\\\\lidar_test\\\\",full.names=T)
   print(read_header("C:\\\\temp\\\\lidar_test\\\\183_302.las"))
   print(read_header(files))

}
\seealso{
\code{\link{read_dtm}}\cr \code{\link{gridmetrics}}\cr
}
\author{
Jacob Strunk <Jacob.strunk@usda.gov>
}
