% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rinex_clip.R
\name{rinex_clip}
\alias{rinex_clip}
\title{clip rinex 3.05 files}
\usage{
rinex_clip(
  dir_rinex = NA,
  dir_out = NA,
  path_rinex = NA,
  path_out = NA,
  start_epoch = 1,
  finish_epoch = NA,
  ignore.short.file = T,
  version_rnx = "3.05"
)
}
\arguments{
\item{dir_rinex}{string, directory of input rinex files}

\item{dir_out}{string, where to send clipped rinex files}

\item{path_rinex}{full path to single or multiple rinex file}

\item{path_out}{full path to single or multiple output rinex file}

\item{start_epoch}{integer, which epoch to start clip}

\item{finish_epoch}{integer, which epoch to finish clip}

\item{ignore.short.file}{T/F some files have fewer epochs than we wish to clip, should this be ignored?}

\item{version_rnx}{"3.05" only rinex version 3.05 is currently supported - or perhaps other versions where epochs are indicated with ">" symbol}
}
\value{
<Delete and Replace>
}
\description{
clip rinex 3.05 files based on starting and ending epoch
}
\details{
Clip rinex files based on epoch number - for example in a 1-second epoch file you
 could clip epochs 30 to 60 from a rinex file. It does not consider the time interval
 you must already new that e.g., a file has 20 second epochs and you want 30 of them to
 represent 10 minutes...

 This program is free software but it is provided WITHOUT WARRANTY
 and with ABSOLUTELY NO GUARANTEE of fitness or functionality for any purpose;
 you can redistribute it and/or modify it under the terms of the GNU
 General Public License as published by the Free Software Foundation;
 either version 2 of the License, or (at your option) any later version.

\cr
Revision History
\tabular{ll}{
1.0 \tab 8/15/2022 Create \cr
1.1 \tab date and revisions.. \cr
}
}
\examples{
 <Delete and Replace>


}
\author{
Jacob Strunk <strunky@gmail.com>
}
