% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rinex_header_edit.r
\name{rinex_header_edit}
\alias{rinex_header_edit}
\title{edit rinex 3.05 header keyword values}
\usage{
rinex_header_edit(
  path_rinex = NA,
  path_out = NA,
  edits = list(c(id = "MARKER NAME", value = "some_project")),
  adds = list(c(id = "MARKER NUMBER", value = "1")),
  version_rnx = "3.05",
  rinex_txt = NA,
  return_rnx = F
)
}
\arguments{
\item{path_rinex}{full path to single or multiple rinex file}

\item{path_out}{full path to single or multiple output rinex file}

\item{edits}{list of named vectors list(c(id="", value=""), c(id="", value="")) where id must exactly match a rinex header keyword and value is an appropriate value}

\item{version_rnx}{"3.05" only rinex version 3.05 is currently supported - or perhaps other versions where epochs are indicated with ">" symbol}

\item{rinex_txt}{instead of providing path, provide rinex text directly}

\item{return_rnx}{T/F should edited filed be returned as text}

\item{add}{list of named vectors list(c(id="", value=""), c(id="", value="")) where id must exactly match a rinex header keyword and value is an appropriate value}
}
\value{
<Delete and Replace>
}
\description{
edit rinex 3.05 header keyword values
}
\details{
This program is free software but it is provided WITHOUT WARRANTY
 and with ABSOLUTELY NO GUARANTEE of fitness or functionality for any purpose;
 you can redistribute it and/or modify it under the terms of the GNU
 General Public License as published by the Free Software Foundation;
 either version 2 of the License, or (at your option) any later version.

\cr
Revision History
\tabular{ll}{
1.0 \tab 8/15/2022 Created \cr
1.1 \tab date and revisions.. \cr
}
}
\examples{
 <Delete and Replace>


}
\author{
Jacob Strunk <strunky@gmail.com>
}
