% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rinex_thin.R
\name{rinex_thin}
\alias{rinex_thin}
\title{Thin temporal density in rinex file}
\usage{
rinex_thin(
  dir_rinex = NA,
  dir_out = NA,
  path_rinex = NA,
  path_out = NA,
  start_hz = 1,
  target_hz = 1/30,
  start_epoch = 1,
  finish_epoch = NA,
  ignore.short.file = T,
  version_rnx = "3.05",
  return_rinex = F
)
}
\arguments{
\item{dir_rinex}{?}

\item{dir_out}{?}

\item{path_rinex}{?}

\item{path_out}{?}

\item{start_hz}{?}

\item{target_hz}{?}

\item{start_epoch}{?}

\item{finish_epoch}{?}

\item{ignore.short.file}{?}

\item{version_rnx}{?}

\item{return_rinex}{?}
}
\value{
<Delete and Replace>
}
\description{
<Thin temporal density in rinex file
}
\details{
This program is free software but it is provided WITHOUT WARRANTY
 and with ABSOLUTELY NO GUARANTEE of fitness or functionality for any purpose;
 you can redistribute it and/or modify it under the terms of the GNU
 General Public License as published by the Free Software Foundation;
 either version 2 of the License, or (at your option) any later version.

\cr
Revision History
\tabular{ll}{
1.0 \tab date and revisions.. \cr
1.1 \tab date and revisions.. \cr
}
}
\examples{
 <Delete and Replace>


}
\seealso{
\code{\link{rinex_header_edit}}\cr \code{\link{rinex_clip}}\cr
}
\author{
Jacob Strunk <strunky@gmail.com>
}
