% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/run_gridmetrics.R
\name{run_gridmetrics}
\alias{run_gridmetrics}
\title{run gridmetrics across a project}
\usage{
run_gridmetrics(
  proj_gpkg_path = NA,
  layer_proj_polys = "RSForInvt_prj",
  layer_proj_config = "RSForInvt_config",
  dir_out = "c:/dir_proc/test_project/gridmetrics",
  n_core = 4,
  gridmetrics_path = "c:\\\\fusion\\\\gridmetrics64.exe",
  heightbreak = 3,
  minht = NA,
  cellsize = NA,
  first = T,
  intensity = F,
  outlier = c(-5, 400),
  fusion_switches = "/nointensity /failnoz",
  xmn = NA,
  xmx = NA,
  ymn = NA,
  ymx = NA,
  new_dtm_path = c(from = NA, to = NA),
  new_las_path = list(from = NA, to = c(NA, NA), prop = c(0.6, 0.4)),
  skip_existing = T,
  subset_tiles = NA,
  existing_coms = c(NA,
    "C:\\\\dir_proc\\\\run_gridmetrics\\\\2017Aug17_100740\\\\all_commands.txt"),
  do_run = T,
  debug = F,
  ...
)
}
\arguments{
\item{n_core}{number of corest to run process on}

\item{gridmetrics_path}{where is gridmetrics.exe (FUSION)}

\item{heightbreak}{Height break for cover calculation}

\item{minht}{set minht for gridmetrics.ex}

\item{cellsize}{override the cellsize set in project create}

\item{first}{T/F use only first returns}

\item{intensity}{T/F include intensity metrics}

\item{outlier}{c(-5,500) range of inclusion}

\item{fusion_switches}{other fusion switches as a string e.g. "/noground"}

\item{xmn, xmx, ymn, ymx}{set extent for analysis}

\item{new_dtm_path}{= c(from = NA, to = NA) rename input file paths (e.g. files have moved)}

\item{new_las_path}{= c(from = NA, to = NA) rename input file paths (e.g. files have moved)}

\item{skip_existing}{skip tiles if they already have output csvs}

\item{existing_coms}{path to existing batch comamnds, incase processing was interrupted the first time}

\item{project_gpkg}{polygon file of intersections created by project_make() function}

\item{table}{output folder name}
}
\value{
<Delete and Replace>
}
\description{
<Delete and Replace>
}
\details{
<Delete and Replace>

\cr
Revision History
\tabular{ll}{
1.0 \tab 2018-01-28 Header added \cr
1.1 \tab 2020-03-17 force csv or geopackage \cr
1.2 \tab 2021-04-30 remove fast cache and stub for processing with lidR \cr
1.3 \tab 2021-07-12 split process across multiple input disk \cr
}
}
\examples{

 gmi=run_gridmetrics(
 project_gpkg="D:\\\\projects\\\\2017_WA_DSM_Pilot_usgs\\\\2017Aug_NAIP_usgs\\\\RSForInvt_project003.gpkg"
,dir_out="I:\\\\projects\\\\2017_WA_DSM_Pilot\\\\2017Aug_NAIP_usgs\\\\gridmetrics_07\\\\"
,dir_dtm="c:\\\\usgs_dtms\\\\dtms\\\\"
,dir_las="D:\\\\naip_2015_laz\\\\"
,n_core=10
,existing_coms="C:\\\\dir_proc\\\\run_gridmetrics\\\\2018Jan21_152618\\\\all_commands.txt"
,n_cache=400
)


}
\seealso{
\code{\link{project_make}}\cr \code{\link{lidR::gridmetrics}}\cr

Desired updates
1. fix issues with mismatched tile size and cell size
2. allow dynamic update of tile sizes
}
\author{
Jacob Strunk <Jstrunk@fs.fed.us>
}
