% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/run_treeseg.R
\name{run_treeseg}
\alias{run_treeseg}
\title{runs treeSeg.exe (USDA FUSION) across a bunch of .dtm caanopy surfaces}
\usage{
run_treeseg(
  dir_dtms = NA,
  dtm_paths = NA,
  dir_out = NA,
  ht_threshold = 40,
  switches = "/shape",
  treeseg_path = "c:\\\\fusion\\\\treeseg.exe",
  ncore = 4
)
}
\arguments{
\item{dir_dtms}{(character - optional) must provide either dir_dtm - a directory with .dtm CSM files - or dtm_paths}

\item{dtm_paths}{(character - optional) must provide either dtm_paths - a vector of .dtm CSM paths - or dir_dtm}

\item{dir_out}{(character - required) where to send outputs}

\item{ht_threshold}{(numeric or character - required) minimum height to be segmented}

\item{switches}{(character - optional) switches to treeseg.exe, see fusion documentation}

\item{treeseg_path}{(character - required) where is the .exe located}

\item{ncore}{(integer - optional) number of cores to use for processing, defaults to 4}
}
\value{
<Delete and Replace>
}
\description{
runs treeSeg.exe  (USDA FUSION) across a bunch of .dtm caanopy surfaces
}
\details{
runs treeSeg.exe  (USDA FUSION) across a bunch of .dtm caanopy surfaces

\cr
Revision History
\tabular{ll}{
1.0 \tab 5/8/2020 FUnction created \cr
}
}
\examples{

run_treeseg(
   dir_dtms = "D:\\\\Box\\\\VMARS\\\\Projects\\\\DAP_evaluation_Meston\\\\R\\\\DAP_Lidar_analysis\\\\RSForInvt\\\\canopymodel\\\\individual"
   ,dtm_paths = NA
   ,dir_out = "D:\\\\Box\\\\VMARS\\\\Projects\\\\DAP_evaluation_Meston\\\\R\\\\DAP_Lidar_analysis\\\\RSForInvt\\\\ITD"
   ,switches = "/shape"
   ,treeseg_path = "c:\\\\fusion\\\\treeseg.exe"
   ,ncore = 4
 )

}
\seealso{
\code{\link{run_canopyModel}}\cr \code{\link{project_create}}\cr
}
\author{
Jacob Strunk <Jacob.strunk@usda.gov>
}
