% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scan_las.R
\name{scan_las}
\alias{scan_las}
\title{scan las files and add them to summary table}
\usage{
scan_las(
  project = "some_project",
  project_year = "2099",
  wkt2 = NA,
  dir_las = "",
  dir_out = paste0(dir_las, "/manage_las/"),
  recursive = F,
  pattern = "[.]la(s|z)$",
  notes = "",
  create_polys = T,
  return = F
)
}
\arguments{
\item{project}{name of lidar project}

\item{project_year}{year of lidar project}

\item{wkt2}{wkt2 string}

\item{dir_las}{where to find lidar files}

\item{pattern}{pattern to use in searching for las files}

\item{notes}{and descriptionn that may be helpful in using a project}

\item{create_polys}{output shapefiles of polygon bboxes}

\item{return}{T/F return objects}
}
\value{
NULL

 or

 list(
  project_id - data.frame with one one row
  ,las_ids - data.frame with as many rows as las files
  ,plys - SpatialPolygonsDataFrame with all of las tile extents
 )
}
\description{
scan las files and add them to summary table
}
\details{
scan las files and add them to summary table

\cr
Revision History
\tabular{ll}{
1.0 \tab 2017 March 08 Created\cr
}
}
\examples{

scan_las(project="test1", project_year="2015",dir_las="C:\\\\temp\\\\lidar_test\\\\",con=con_inv)

}
\seealso{
\code{\link{scan_dtm}}\cr \code{\link{read_las}}\cr
}
\author{
Jacob Strunk <Jstrunk@fs.fed.us>
}
