% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sqlite_to_raster.R
\name{sqlite_to_raster}
\alias{sqlite_to_raster}
\alias{colsAll}
\alias{colsID}
\alias{colsAllX}
\alias{colsSomeX}
\alias{colsFewX}
\title{helpers to convert sqlite table to rasters}
\usage{
sqlite_to_raster(
  db,
  tb_gm = "gm",
  colsxy = c("center_x", "center_y"),
  cols2Raster = colsSomeX(),
  res = c(NA, NA),
  format = ".img",
  dirOut = "E:\\\\projects\\\\2017_NAIP\\\\rasters\\\\",
  raster_prefix = "",
  wkt2 = NA,
  nProc = 4,
  doDebug = F,
  doBuild = F,
  set9999 = c(0, NA, -9999),
  big_raster = F
)

colsAll(thresh = "6_00")

colsID()

colsAllX(thresh = "6_00")

colsSomeX(thresh = "6_00")

colsFewX(thresh = "6_00")
}
\arguments{
\item{db}{a database object, tested for sqlite database}

\item{tb_gm}{which table has gridmetrics data}

\item{colsxy}{names of xy columns}

\item{cols2Raster}{which columns to grab}

\item{format}{raster formate  e.g. .tif , .img etc}

\item{dirOut}{where to export rasters}

\item{raster_prefix}{give a name to rasters for id purposes}

\item{nProc}{number of cores to use in converting to raster}

\item{doDebug}{run in debug mode and only read 50k rows?}

\item{thresh}{character string like "2_50" that matches height representation in column names in tb_gm}

\item{crs}{proj4 string or other crs notation used by raster package (see writeRaster)}
}
\value{
<Delete and Replace>
}
\description{
<Delete and Replace>
}
\details{
<Delete and Replace>

\cr
Revision History
\tabular{ll}{
1.0 \tab 3/16/2020 Function created \cr
}
}
\examples{
 <Delete and Replace>

}
\seealso{
\code{\link{rasterFromXYZ}}\cr \code{\link{writeRaster}}\cr
}
\author{
Jacob Strunk <Jstrunk@fs.fed.us>
}
