% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bbox2polys.R
\name{bbox2polys}
\alias{bbox2polys}
\title{create polygons from dataframe of bounding coordinates and ids}
\usage{
bbox2polys(
  df_ext,
  cols_ext = c(xmin = "min_x", xmax = "max_x", ymin = "min_y", ymax = "max_y"),
  wkt2 = NA
)
}
\arguments{
\item{df_ext}{a data.frame with 4 columns of coordinates: minx, maxx, miny, and maxy indicating bounding boxes or extents}

\item{cols_ext}{a named vector with names 'xmin','xmax','ymin','ymax' indicating where bound box coordinates are stored}

\item{wkt2}{standard crs format for sf and terra packages}
}
\value{
sf spatial data frame with all original fields and added geometry field
}
\description{
create polygons from dataframe of bounding coordinates and ids
}
\details{
create polygons from dataframe of bounding coordinates and ids


 This program is free software but it is provided WITHOUT WARRANTY
 and with ABSOLUTELY NO GUARANTEE of fitness or functionality for any purpose;
 you can redistribute it and/or modify it under the terms of the GNU
 General Public License as published by the Free Software Foundation;
 either version 2 of the License, or (at your option) any later version.

\cr
Revision History
\tabular{ll}{
1.0 \tab 2017 March 08 Created\cr
}
}
\examples{
 #none yet

}
\seealso{
\code{\link{scan_las}}\cr \code{\link{read_las}}\cr
}
\author{
Jacob Strunk <strunky@gmail.com>
}
