% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/project_create.R
\name{project_create}
\alias{project_create}
\title{Build point cloud processing tiling project (.laz, .las, .dtm) for use with FUSION}
\usage{
project_create(
  dir_las = NA,
  dir_dtm = NA,
  recurse_dtm = F,
  recurse_las = F,
  dir_out = "c:/lidar_projects/",
  project_name = "RSForTools",
  layer_processing_tiles = "processing_tiles",
  layer_configuration = "project_settings",
  overwrite_project = T,
  project_dtm = "dtm",
  project_las = "las",
  dtm_year = "2099",
  las_year = "2099",
  do_scan_dtms = T,
  do_scan_las = T,
  duplicate_las = c("ignore", "remove"),
  duplicate_dtm = c("ignore", "remove"),
  tile_size = 1650,
  pixel_size = 66,
  xmn = c(NA, 561066),
  xmx = c(NA, 2805066),
  ymn = c(NA, 33066),
  ymx = c(NA, 1551066),
  wkt2 = c(NA,
    "+proj=lcc +lat_1=47.33333333333334 +lat_2=45.83333333333334 +lat_0=45.33333333333334 +lon_0=-120.5 +x_0=500000.0001016001 +y_0=0 +ellps=GRS80 +towgs84=0,0,0,0,0,0,0 +units=us-ft +no_defs"),
  mask = NA,
  return = T
)
}
\arguments{
\item{dir_las}{where are las files}

\item{dir_dtm}{where are FUSION dtm files - eventuall enable any dtm type (.img, .tif etc)}

\item{recurse_dtm}{recurse into las folder subdirectories}

\item{recurse_las}{recurse into dtm folder subdirectories}

\item{dir_out}{where to place project}

\item{project_name}{project name}

\item{layer_processing_tiles}{name of layer in geopackage to hold processing tile geometry}

\item{layer_configuration}{name of layer in geopackage to hold project configuration informaton}

\item{overwrite_project}{T/F should the original project be overwritten}

\item{project_dtm}{dtm project name}

\item{project_las}{las project name}

\item{dtm_year}{year of dtm}

\item{las_year}{year of las}

\item{do_scan_dtms}{T/F re-scan dtm files}

\item{do_scan_las}{T/F re-scan las files}

\item{duplicate_las}{c("ignore","remove") - how to handle las tiles with same base names in different directories}

\item{duplicate_dtm}{c("ignore","remove") - how to handle dtm tiles with same base names in different directories}

\item{tile_size}{processing tile size}

\item{pixel_size}{raster pixel size}

\item{xmn, xmx, ymn, ymx}{bound box for processing grid}

\item{wkt2}{CRS projection string in wkt2 format}

\item{mask}{sf vector layer that can serve as a mask for analysis}

\item{return}{T/F return project as an sf object}
}
\value{
<Delete and Replace>
}
\description{
Scans lidar and dems and finds out where they intersect a project tiling scheme
}
\details{
<Delete and Replace>

\cr
Revision History
\tabular{ll}{
1.0 \tab 5/07/2020 New package derived from old lasR package \cr
1.1 \tab 6/03/2024 Various updates including roxygen header \cr
}
}
\examples{

proj_tn = RSForInvt::project_create(
  #'proj = project_create(
  dir_las="D:\\\\Box\\\\VMARS\\\\Projects\\\\DAP_evaluation_Meston\\\\Data\\\\Tennessee\\\\lidar_tiles\\\\"
  ,dir_dtm="D:\\\\Box\\\\VMARS\\\\Projects\\\\DAP_evaluation_Meston\\\\Data\\\\Tennessee\\\\DTM_fusion\\\\"
  ,path_gpkg_out="D:\\\\Box\\\\VMARS\\\\Projects\\\\DAP_evaluation_Meston\\\\R\\\\DAP_Lidar_analysis\\\\RSForInvt\\\\project\\\\TNLidar_RSForInvtProject.gpkg"
  ,project_name = "RSForInvt_prj"
  ,layer_processing_tiles = "RSForInvt_config"
  ,overwrite_project = T
  ,project_dtm="lidar_dtm"
  ,project_las="lidar_las"
  ,dtm_year="2018"
  ,las_year="2018"
  ,do_scan_dtms=F #'we already scanned the dtm folder - ok, to rescan, but slower
  ,do_scan_las=F #'we already scanned the las folder - ok, to rescan, but slower
  ,duplicate_las = c("ignore","remove")
  ,duplicate_dtm = c("ignore","remove")
  ,tile_size=1650
  ,pixel_size=66
  #',xmn=561066,xmx=2805066,ymn=33066,ymx=1551066
  ,proj4 = "+proj=lcc +lat_1=47.33333333333334 +lat_2=45.83333333333334 +lat_0=45.33333333333334 +lon_0=-120.5 +x_0=500000.0001016001 +y_0=0 +ellps=GRS80 +towgs84=0,0,0,0,0,0,0 +units=us-ft +no_defs"
  ,mask=NA
  ,return=T
)


}
\seealso{
\code{\link{scan_dtm}}\cr \code{\link{scan_las}}\cr


desired updates:
 add ability to scan fusion dtms and generic raster dtms
}
\author{
Jacob Strunk <strunky@gmail.com>
}
