% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scan_dtm.R
\name{scan_dtm}
\alias{scan_dtm}
\title{scan a directory of dtm files}
\usage{
scan_dtm(
  project = "wa_dtm",
  project_year = "2099",
  proj4_name = c(NA, "NAD_1983_HARN_StatePlane_Washington_South_FIPS_4601_Feet"),
  proj4 = NA,
  dir_dtm = "",
  recursive = F,
  pattern = "[.]dtm$",
  notes = "",
  create_polys = T,
  update = T,
  return = F,
  debug = F
)
}
\arguments{
\item{project}{name of lidar project}

\item{project_year}{year of lidar project}

\item{proj4_name}{human interpretable name of projection if known}

\item{proj4}{override proj4 string associated with DTMS}

\item{dir_dtm}{where to find dtm files}

\item{recursive}{T/F recurse into dir_dtm subdirectories ?}

\item{pattern}{pattern to use in searching for dtm files}

\item{notes}{and descriptionn that may be helpful in using a project}

\item{create_polys}{output shapefiles of polygon bboxes}

\item{return}{T/F return objects}
}
\value{
NULL

 or

 list(
  project_id - data.frame with one one row
  ,dtm_ids - data.frame with as many rows as dtms
  ,plys - SpatialPolygonsDataFrame with all of dtm extents
 )
}
\description{
scan a directory of dtm files
}
\details{
scan a directory of dtm files

\cr
Revision History
\tabular{ll}{
1.0 \tab 2017 March 08 Created\cr
}
}
\examples{
 <Delete and Replace>

}
\seealso{
\code{\link{scan_las}}\cr \code{\link{read_dtm}}\cr \code{\link{read_dtm_header}}\cr
}
\author{
Jacob Strunk <Jstrunk@fs.fed.us>
}
