% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sqlite_to_raster.R
\name{sqlite_to_raster}
\alias{sqlite_to_raster}
\alias{colsAll}
\alias{colsID}
\alias{colsAllX}
\alias{colsSomeX}
\alias{colsFewX}
\title{helpers to convert sqlite table to rasters}
\usage{
sqlite_to_raster(
  db,
  tb_gm = "gm",
  colsxy = c("center_x", "center_y"),
  cols2Raster = colsSomeX(),
  res = c(NA, NA),
  format = ".img",
  dirOut = "E:\\\\projects\\\\2017_NAIP\\\\rasters\\\\",
  raster_prefix = "",
  wkt2 = NA,
  nproc = 4,
  doDebug = F,
  debugRows = 5e+05,
  set9999 = c(0, NA, -9999),
  sfmask = NA
)

colsAll(thresh = "6_00")

colsID()

colsAllX(thresh = "6_00")

colsSomeX(thresh = "6_00")

colsFewX(thresh = "6_00")
}
\arguments{
\item{db}{a database object, tested for sqlite database}

\item{tb_gm}{which table has gridmetrics data}

\item{colsxy}{names of xy columns}

\item{cols2Raster}{which columns to grab}

\item{format}{raster formate  e.g. .tif , .img etc}

\item{dirOut}{where to export rasters}

\item{raster_prefix}{give a name to rasters for id purposes}

\item{wkt2}{crs approach used by terra or sf,}

\item{nproc}{number of cores to use in converting to raster}

\item{doDebug}{run in debug mode and only read 50k rows?}

\item{thresh}{character string like "2_50" that matches height representation in column names in tb_gm}

\item{debugRow}{number of rows to test / debug on}
}
\value{
<Delete and Replace>
}
\description{
<Delete and Replace>
}
\details{
<Delete and Replace>

\cr
Revision History
\tabular{ll}{
1.0 \tab 3/16/2020 Function created \cr
1.0 \tab 8/16/2024 remove col_test argument - now handled while reading csvs \cr
}
}
\examples{

  dir_sqlite = "D:/temp/OR_dap_gridmetrics_sqlite/"
  nms_x1b = c('ht_p90','percentage_all_returns_above_3_00','profile_area','ht_mode','ht_stddev','ht_p05','ht_p10','ht_p50','canopy_relief_ratio','ht_mean')
  #function not compiled yet, from lasR package, have to source it..
  sqlite_to_raster(file.path(dir_sqlite,"OR2022_NAIP_Metrics.db")
                   #,cols2Raster = c("ht_p90", "ht_p30","canopy_relief_ratio","percentage_first_returns_above_6_00")
                   , dirOut = "d:/temp/or_dap_2022_gridmetrics_raster/"
                   , wkt2 = terra::crs("EPSG:6557")
                   , cols2Raster = nms_x1b
                   , nproc = 5
                   , doDebug=T
                   )


}
\seealso{
\code{\link{rasterFromXYZ}}\cr \code{\link{writeRaster}}\cr
}
\author{
Jacob Strunk <Jstrunk@fs.fed.us>
}
